add support for recipes

+++ graphql/resolvers/ingredients.js
@@ -4,25 +4,34 @@ const { Ingredient } = require("../../models");
 
 module.exports = {
   Mutation: {
-    async deleteIngredient(_, { deleteIngredientInput: { name, userId } }) {
-      Ingredient.destroy({
-        where: {
-          name,
-          userId,
-        },
-      });
+    async deleteIngredient(
+      _,
+      { deleteIngredientInput: { ingredientsToDelete, userId } }
+    ) {
+      ingredientsToDelete
+        .split(",")
+        .forEach(
+          async (item) =>
+            await Ingredient.destroy({ where: { name: item.trim(), userId } })
+        );
 
-      return "ok";
+      return `ingredients deleted: ${ingredientsToDelete}`;
     },
     async editUserIngredients(
       _,
       { editUserIngredientsInput: { input, userId } }
     ) {
-      input
-        .split(",")
-        .forEach(
-          async (item) => await Ingredient.create({ name: item.trim(), userId })
-        );
+      const ingredientsList = await Ingredient.findAll({
+        where: {
+          userId: userId,
+        },
+      });
+
+      ingredientsList.forEach(() => Ingredient.destroy({ where: { userId } }));
+
+      input.forEach(
+        async (item) => await Ingredient.create({ name: item.trim(), userId })
+      );
 
       const updatedIngredientsList = await Ingredient.findAll({
         where: {
